fix: remove route.ts and move function
